requirements.txt
aiohttp==3.8.4
discord.py==2.3.2
PyNaCl==1.5.0
python-dotenv==1.0.0
validators==0.20.0
yt-dlp==2023.7.6
pydantic-settings==2.0.2